refactor(report): move evaluation types and report generator to report directory
Restructure the codebase by relocating evaluation-related types and the report generator to a dedicated 'report' directory. This improves maintainability and logical organization of the code.
